imagebind_LLM/requirements.txt
fairscale
sentencepiece
Pillow
opencv-python
gradio
tqdm
autofaiss